Ask user before opening editor to edit snippet
Previously if a snippet was a single line, there was no way of opening
it in an external editor. This meant (among other things) that is was
not possible to add a line break to a single line snippet.

It also meant for multi line snippets that the editor had to be opened
and closed in order for the edit to be completed even if the user only
wanted to edit tags or description.

This is slightly more involved when making simple edits to single lines
but the consistency across all snippets makes it clearer how things
work.

Fixes #104